Fixing compilation issues with int - size_t comparisons

+++ inc/TRestDetectorReadoutPlane.h
@@ -118,7 +118,7 @@ class TRestDetectorReadoutPlane {
     TRestDetectorReadoutModule& operator[](int mod) { return fReadoutModules[mod]; }
 
     /// Returns a pointer to a readout module using its std::vector index
-    TRestDetectorReadoutModule* GetModule(int mod) {
+    TRestDetectorReadoutModule* GetModule(size_t mod) {
         if (mod >= GetNumberOfModules()) return nullptr;
         return &fReadoutModules[mod];
     }
